+++ 2024/src/03/part_1.zig
@@ -6,9 +6,9 @@ const pretty = @import("pretty");
 const debug = @import("std").debug;
 const Regex = @import("regex").Regex;
 
-pub fn run(allocator: std.mem.Allocator, input: parse.Input) !u32 {
+pub fn run(allocator: std.mem.Allocator, input: parse.Input) !u64 {
     _ = allocator;
-    var total: u32 = 0;
+    var total: u64 = 0;
 
     for (input.instructions) |instruction| {
         total += instruction.a * instruction.b;
