+++ 2024/src/05/part_2.zig
@@ -6,7 +6,7 @@ const debug = @import("common").debug;
 const part_1 = @import("part_1.zig");
 
 const Rules = struct {
-    map: std.AutoHashMap(u32, []u32),
+    map: std.AutoHashMap(u64, []u64),
     allocator: std.mem.Allocator,
 
     pub fn deinit(self: *Rules) void {
@@ -18,10 +18,10 @@ const Rules = struct {
     }
 };
 
-pub fn run(allocator: std.mem.Allocator, input: parse.Input) !u32 {
+pub fn run(allocator: std.mem.Allocator, input: parse.Input) !u64 {
     const updates = input.updates;
 
-    var invalid_updates = try std.ArrayList([]u32).initCapacity(allocator, updates.len);
+    var invalid_updates = try std.ArrayList([]u64).initCapacity(allocator, updates.len);
     defer invalid_updates.deinit();
 
     for (updates) |update| {
@@ -48,16 +48,16 @@ pub fn run(allocator: std.mem.Allocator, input: parse.Input) !u32 {
 // e.g. if A -> B, B -> C, C -> D, we deduct A -> D, but if B is not in the update, we should not deduct A -> D
 fn removeUnusedRules(
     allocator: std.mem.Allocator,
-    left_before_rights: std.AutoHashMap(u32, []u32),
-    update: []u32,
+    left_before_rights: std.AutoHashMap(u64, []u64),
+    update: []u64,
 ) !Rules {
-    var used_pages = std.AutoHashMap(u32, void).init(allocator);
+    var used_pages = std.AutoHashMap(u64, void).init(allocator);
     defer used_pages.deinit();
     for (update) |page| {
         try used_pages.put(page, {});
     }
 
-    var result = std.AutoHashMap(u32, []u32).init(allocator);
+    var result = std.AutoHashMap(u64, []u64).init(allocator);
 
     var it = left_before_rights.iterator();
     while (it.next()) |entry| {
@@ -75,7 +75,7 @@ fn removeUnusedRules(
         }
 
         if (count > 0) {
-            var new_rights = try allocator.alloc(u32, count);
+            var new_rights = try allocator.alloc(u64, count);
             var i: usize = 0;
             for (rights) |right| {
                 if (used_pages.get(right) != null) {
@@ -95,13 +95,13 @@ fn removeUnusedRules(
 // expands rules transitive, e.g. if A -> B, B -> C, we add A -> C
 fn expandRulesTransitive(
     allocator: std.mem.Allocator,
-    left_before_rights: std.AutoHashMap(u32, []u32),
+    left_before_rights: std.AutoHashMap(u64, []u64),
 ) !Rules {
-    var result = std.AutoHashMap(u32, []u32).init(allocator);
+    var result = std.AutoHashMap(u64, []u64).init(allocator);
 
     var it = left_before_rights.iterator();
     while (it.next()) |entry| {
-        const rights = try allocator.dupe(u32, entry.value_ptr.*);
+        const rights = try allocator.dupe(u64, entry.value_ptr.*);
         try result.put(entry.key_ptr.*, rights);
     }
 
@@ -109,7 +109,7 @@ fn expandRulesTransitive(
     while (changed) {
         changed = false;
 
-        var updates = std.ArrayList(struct { key: u32, value: []u32 }).init(allocator);
+        var updates = std.ArrayList(struct { key: u64, value: []u64 }).init(allocator);
         defer updates.deinit();
 
         var outer_it = result.iterator();
@@ -144,8 +144,8 @@ fn expandRulesTransitive(
     };
 }
 
-fn addUniqueRights(allocator: std.mem.Allocator, current: []u32, to_add: []u32) ![]u32 {
-    var set = std.AutoHashMap(u32, void).init(allocator);
+fn addUniqueRights(allocator: std.mem.Allocator, current: []u64, to_add: []u64) ![]u64 {
+    var set = std.AutoHashMap(u64, void).init(allocator);
     defer set.deinit();
 
     for (current) |value| {
@@ -156,7 +156,7 @@ fn addUniqueRights(allocator: std.mem.Allocator, current: []u32, to_add: []u32) 
         try set.put(value, {});
     }
 
-    var result = try allocator.alloc(u32, set.count());
+    var result = try allocator.alloc(u64, set.count());
     var i: usize = 0;
     var it = set.keyIterator();
     while (it.next()) |key| {
@@ -169,13 +169,13 @@ fn addUniqueRights(allocator: std.mem.Allocator, current: []u32, to_add: []u32) 
 
 fn fixUpdate(
     allocator: std.mem.Allocator,
-    update: []u32,
-    left_before_rights: std.AutoHashMap(u32, []u32),
+    update: []u64,
+    left_before_rights: std.AutoHashMap(u64, []u64),
 ) !void {
     var changes_made = true;
     while (changes_made) {
         changes_made = false;
-        var past_pages = std.AutoHashMap(u32, bool).init(allocator);
+        var past_pages = std.AutoHashMap(u64, bool).init(allocator);
         defer past_pages.deinit();
 
         var right_index: usize = 0;
@@ -201,7 +201,7 @@ fn fixUpdate(
     }
 }
 
-fn findIndex(update: []u32, value: u32) usize {
+fn findIndex(update: []u64, value: u64) usize {
     var i: usize = 0;
     while (i < update.len) {
         if (update[i] == value) {
@@ -212,7 +212,7 @@ fn findIndex(update: []u32, value: u32) usize {
     return std.math.maxInt(usize);
 }
 
-fn bubbleUp(update: []u32, left_index: usize, right_index: usize) void {
+fn bubbleUp(update: []u64, left_index: usize, right_index: usize) void {
     const value = update[right_index];
     var i = right_index;
     while (i > left_index) {
