+++ 2024/src/06/part_2_test.zig
@@ -3,7 +3,7 @@ const testing = std.testing;
 const parse = @import("parse.zig");
 const part_2 = @import("part_2.zig");
 
-const EXAMPLE_ANSWER_2: u32 = 6;
+const EXAMPLE_ANSWER_2: u64 = 6;
 
 test "part 2 example" {
     var gpa = std.heap.GeneralPurposeAllocator(.{}){};
@@ -15,4 +15,3 @@ test "part 2 example" {
     const result = part_2.run(allocator, input);
     try testing.expectEqual(EXAMPLE_ANSWER_2, result);
 }
-
